Inlcude ZooKeeper serverset member name in endpoint comparisons (#41)
## Description

Using the ZooKeeper ServerSetSource, if service instances crash and restart on the same host:port it can lead to all endpoints for the service being removed and a total outage to clients, due to the following sequence of events:

* A service instance crashes and does not cleanly remove itself from the serverset, leaving a "zombie" member.
* A service instance starts on the same host:port and registers itself in the serverset as a new member.
* ServerSetSource is notified of the new serverset member, but ignores it because that host:port is already present in the same source.
* The ZK node of the crashed service instance is deleted, presumably because its ZK session expires.
* ServerSetSource is notified of the removed serverset member. It removes the endpoint, leaving no endpoints for the service at all.

This has shown up with this sequence of log events:

```
2024-07-09 19:01:23,323 [INFO] tellapart.aurproxy.source.source: Adding endpoint: 10.52.64.35:80
2024-07-10 02:03:47,672 [INFO] tellapart.aurproxy.source.source: Not adding endpoint - already known. 10.52.64.35:80
2024-07-10 02:04:05,715 [INFO] tellapart.aurproxy.source.source: Removing endpoint: 10.52.64.35:80
2024-07-10 13:46:56,305 [INFO] tellapart.aurproxy.source.source: Not removing endpoint - not known. 10.52.64.35:80
```

## Changes

The problem is that the ProxySource logic endpoint objects adds Endpoint objects to a set:

https://github.com/amperity/aurproxy/blob/d97261f56370886c8302071d9e66587284160f90/tellapart/aurproxy/source/source.py#L56-L86

however Endpoints are only compared based on host:port:

https://github.com/amperity/aurproxy/blob/d97261f56370886c8302071d9e66587284160f90/tellapart/aurproxy/config/endpoint.py#L63-L74

To fix the issue we can include a unique key in that comparison that distinguishes the old service instance from the new service instance. For a ZK serverset we can include the ZK node name, also know as the serverset member name. If members are created as sequential nodes then ZK guarantees their names are unique, with monotonically increasing suffixes (https://zookeeper.apache.org/doc/r3.5.4-beta/zookeeperProgrammers.html#Sequence+Nodes+--+Unique+Naming).

Also added the endpoint context map, which includes the unique key, to logging in ProxySource add and remove logic.